chore: rename components and llmEvaluation, update imports
