+++ agents/prompt-evaluator/components/PointDescription.tsx
@@ -30,4 +30,3 @@ export default function PointDescription() {
       </div>
     );
   }
-
